+++ SolStandard/Content/TmxMaps/Debug_01.tmx
@@ -1,5 +1,5 @@
 <?xml version="1.0" encoding="UTF-8"?>
-<map version="1.0" tiledversion="1.1.6" orientation="orthogonal" renderorder="right-down" width="32" height="20" tilewidth="32" tileheight="32" infinite="0" nextobjectid="688">
+<map version="1.0" tiledversion="1.1.6" orientation="orthogonal" renderorder="right-down" width="32" height="20" tilewidth="32" tileheight="32" infinite="0" nextobjectid="689">
  <tileset firstgid="1" name="entities-32" tilewidth="32" tileheight="32" tilecount="448" columns="7">
   <image source="F:/Pictures/Assets/PAID/CustomWonder/Terrain/entities-32.png" width="224" height="2048"/>
  </tileset>
@@ -2439,6 +2439,7 @@
   </object>
   <object id="686" name="Throne" type="Decoration" gid="204" x="384" y="256" width="32" height="32"/>
   <object id="687" name="Bridge" type="Movable" gid="3149" x="416" y="480" width="32" height="32"/>
+  <object id="688" name="Push Block" type="Shovable" gid="6053" x="512" y="256" width="32" height="32"/>
  </objectgroup>
  <objectgroup name="Items">
   <object id="657" name="Door LK1 Key" type="Key" gid="388" x="288" y="288" width="32" height="32">
