+++ SolStandard/Entity/General/Artillery.cs
@@ -1,8 +1,8 @@
 ﻿using System.Collections.Generic;
 using Microsoft.Xna.Framework;
 using SolStandard.Entity.Unit;
-using SolStandard.Entity.Unit.Skills;
-using SolStandard.Entity.Unit.Skills.Terrain;
+using SolStandard.Entity.Unit.Actions;
+using SolStandard.Entity.Unit.Actions.Terrain;
 using SolStandard.HUD.Window.Content;
 using SolStandard.Map.Elements;
 using SolStandard.Utility;
