+++ SolStandard/Entity/General/Portal.cs
@@ -2,8 +2,8 @@
 using Microsoft.Xna.Framework;
 using SolStandard.Containers;
 using SolStandard.Entity.Unit;
-using SolStandard.Entity.Unit.Skills;
-using SolStandard.Entity.Unit.Skills.Terrain;
+using SolStandard.Entity.Unit.Actions;
+using SolStandard.Entity.Unit.Actions.Terrain;
 using SolStandard.HUD.Window.Content;
 using SolStandard.Map;
 using SolStandard.Map.Elements;
