+++ SolStandard/Entity/IItem.cs
@@ -1,4 +1,4 @@
-﻿using SolStandard.Entity.Unit.Skills;
+﻿using SolStandard.Entity.Unit.Actions;
 using SolStandard.Utility;
 
 namespace SolStandard.Entity
