+++ SolStandard/Entity/Unit/Actions/BasicAttack.cs
@@ -10,7 +10,7 @@ using SolStandard.Utility;
 using SolStandard.Utility.Assets;
 using SolStandard.Utility.Events;
 
-namespace SolStandard.Entity.Unit.Skills
+namespace SolStandard.Entity.Unit.Actions
 {
     public class BasicAttack : UnitAction
     {
