+++ SolStandard/Entity/Unit/Actions/Champion/Cover.cs
@@ -9,42 +9,41 @@ using SolStandard.Utility;
 using SolStandard.Utility.Assets;
 using SolStandard.Utility.Events;
 
-namespace SolStandard.Entity.Unit.Skills.Archer
+namespace SolStandard.Entity.Unit.Actions.Champion
 {
-    public class Draw : UnitAction
+    public class Cover : UnitAction
     {
         private readonly int statModifier;
         private readonly int duration;
 
-        public Draw(int duration, int statModifier) : base(
-            icon: SkillIconProvider.GetSkillIcon(SkillIcon.Draw, new Vector2(32)),
-            name: "Draw",
-            description: "Increase own attack range by [+" + statModifier + "] for [" + duration + "] turns.",
+        public Cover(int duration, int statModifier) : base(
+            icon: SkillIconProvider.GetSkillIcon(SkillIcon.Cover, new Vector2(32)),
+            name: "Cover",
+            description: "Increase an ally's DEF by [+" + statModifier + "] for [" + duration + "] turns.",
             tileSprite: MapDistanceTile.GetTileSprite(MapDistanceTile.TileType.Action),
-            range: new[] {0}
+            range: new[] {1}
         )
         {
-            //Add one to the duration to compensate for the counter going down right after the user's turn ends.
-            this.duration = duration + 1;
             this.statModifier = statModifier;
+            this.duration = duration;
         }
 
         public override void ExecuteAction(MapSlice targetSlice, GameMapContext gameMapContext, BattleContext battleContext)
         {
             GameUnit targetUnit = UnitSelector.SelectUnit(targetSlice.UnitEntity);
 
-            if (TargetIsSelfInRange(targetSlice, targetUnit))
+            if (TargetIsAnAllyInRange(targetSlice, targetUnit))
             {
                 MapContainer.ClearDynamicAndPreviewGrids();
 
                 Queue<IEvent> eventQueue = new Queue<IEvent>();
-                eventQueue.Enqueue(new CastBuffEvent(ref targetUnit, new AtkRangeStatUp(duration, statModifier)));
+                eventQueue.Enqueue(new CastBuffEvent(ref targetUnit, new DefStatUp(duration, statModifier)));
                 eventQueue.Enqueue(new EndTurnEvent(ref gameMapContext));
                 GlobalEventQueue.QueueEvents(eventQueue);
             }
             else
             {
-                MapContainer.AddNewToastAtMapCursor("Invalid target!", 50);
+                MapContainer.AddNewToastAtMapCursor("Not an ally in range!", 50);
                 AssetManager.WarningSFX.Play();
             }
         }
