+++ SolStandard/Entity/Unit/Actions/Shove.cs
@@ -8,7 +8,7 @@ using SolStandard.Utility;
 using SolStandard.Utility.Assets;
 using SolStandard.Utility.Events;
 
-namespace SolStandard.Entity.Unit.Skills
+namespace SolStandard.Entity.Unit.Actions
 {
     public class Shove : UnitAction
     {
