+++ SolStandard/Entity/Unit/Actions/Terrain/ArtilleryAction.cs
@@ -3,7 +3,7 @@ using SolStandard.Map.Elements;
 using SolStandard.Map.Elements.Cursor;
 using SolStandard.Utility;
 
-namespace SolStandard.Entity.Unit.Skills.Terrain
+namespace SolStandard.Entity.Unit.Actions.Terrain
 {
     public class ArtilleryAction : UnitAction
     {
