--- SolStandard/Entity/Unit/Skills/Terrain/PickUpCurrencyAction.cs
@@ -1,52 +1,0 @@
-﻿using System.Collections.Generic;
-using SolStandard.Containers;
-using SolStandard.Containers.Contexts;
-using SolStandard.Entity.General;
-using SolStandard.Map.Elements;
-using SolStandard.Map.Elements.Cursor;
-using SolStandard.Utility.Assets;
-using SolStandard.Utility.Events;
-
-namespace SolStandard.Entity.Unit.Skills.Terrain
-{
-    public class PickUpCurrencyAction : UnitAction
-    {
-        private readonly Currency currency;
-
-        public PickUpCurrencyAction(Currency currency) : base(
-            icon: currency.RenderSprite,
-            name: "Pick Up",
-            description: "Add [" + currency.Value + "] to your unit's money count.",
-            tileSprite: MapDistanceTile.GetTileSprite(MapDistanceTile.TileType.Action),
-            range: currency.Range
-        )
-        {
-            this.currency = currency;
-        }
-
-        public override void ExecuteAction(MapSlice targetSlice, GameMapContext gameMapContext, BattleContext battleContext)
-        {
-            if (SelectingItemAtUnitLocation(targetSlice))
-            {
-                MapContainer.ClearDynamicAndPreviewGrids();
-
-                Queue<IEvent> eventQueue = new Queue<IEvent>();
-                eventQueue.Enqueue(new PickUpCurrencyEvent(currency));
-                eventQueue.Enqueue(new WaitFramesEvent(10));
-                eventQueue.Enqueue(new EndTurnEvent(ref gameMapContext));
-                GlobalEventQueue.QueueEvents(eventQueue);
-            }
-            else
-            {
-                MapContainer.AddNewToastAtMapCursor("Cannot pick up money here!", 50);
-                AssetManager.WarningSFX.Play();
-            }
-        }
-
-        private bool SelectingItemAtUnitLocation(MapSlice targetSlice)
-        {
-            return currency.MapCoordinates == GameContext.ActiveUnit.UnitEntity.MapCoordinates &&
-                   targetSlice.DynamicEntity != null;
-        }
-    }
-}
