+++ SolStandard/Map/Elements/IActionTile.cs
@@ -1,5 +1,5 @@
 ﻿using Microsoft.Xna.Framework;
-using SolStandard.Entity.Unit.Skills;
+using SolStandard.Entity.Unit.Actions;
 
 namespace SolStandard.Map.Elements
 {
